feature: added GalleryItemToolBar Component
